+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+### Added
+* Expose new `loadEnvVarsFromConfig` function to use instead of `EnvVarLoaderProvider`.
+
+### Changed
+* *Nothing*
+
+### Deprecated
+* Deprecate `EnvVarLoaderProvider`. Use `loadEnvVarsFromConfig` instead.
+
+### Removed
+* *Nothing*
+
+### Fixed
+* *Nothing*
+
+
 ## [3.1.0] - 2024-07-22
 ### Added
 * Support `laminas/laminas-servicemanager` >=4.2