+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [3.2.0] - 2024-10-14
 ### Added
 * Expose new `loadEnvVarsFromConfig` function to use instead of `EnvVarLoaderProvider`.
 * Expose new `formatEnvVarValue` function that stringifies a value to use as an env var.